+++ src/main/java/marquez/service/models/StreamMeta.java
@@ -54,7 +54,7 @@ public final class StreamMeta extends DatasetMeta {
   }
 
   @Override
-  public UUID version(@NonNull NamespaceName namespaceName, @NonNull DatasetName datasetName) {
+  public Version version(@NonNull NamespaceName namespaceName, @NonNull DatasetName datasetName) {
     final byte[] bytes =
         VERSION_JOINER
             .join(
@@ -65,6 +65,6 @@ public final class StreamMeta extends DatasetMeta {
                 schemaLocation.toString(),
                 getFields().stream().map(DatasetMeta::joinField).collect(joining(VERSION_DELIM)))
             .getBytes(UTF_8);
-    return UUID.nameUUIDFromBytes(bytes);
+    return Version.of(UUID.nameUUIDFromBytes(bytes));
   }
 }